Refactor get_mod_roles() to new module
